Adding loginResponse class and debugging cors issues
