+++ src/main/java/com/templateproject/api/config/SecurityConfiguration.java
@@ -49,6 +49,7 @@ public class SecurityConfiguration {
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         return http.csrf(csrf -> csrf.disable())
+                .cors(Customizer.withDefaults())
                 .authorizeHttpRequests(auth -> {
                     auth.requestMatchers(HttpMethod.GET, "/articles/**", "/categories/**",
                             "/swagger-ui/**" ,"/v3/**", "/index.html").permitAll()
