+++ src/main/java/com/templateproject/api/entity/LoginResponse.java
@@ -1,0 +1,28 @@
+package com.templateproject.api.entity;
+
+public class LoginResponse {
+    private String token;
+    private User user;
+
+
+    public LoginResponse(String token, User user) {
+        this.token = token;
+        this.user = user;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+    }
+}
