--- src/main/java/com/templateproject/api/entity/Token.java
@@ -1,18 +1,0 @@
-package com.templateproject.api.entity;
-
-public class Token {
-    private String token;
-
-
-    public Token(String token) {
-        this.token = token;
-    }
-
-    public String getToken() {
-        return token;
-    }
-
-    public void setToken(String token) {
-        this.token = token;
-    }
-}
